refactor(server): extract resume version routes (8/8 — server split done!)
Final extraction of the server.py monolith split (CLAUDE.md tech-debt #2).

Routes moved (5):
- GET    /api/resume/versions
- POST   /api/resume/versions
- POST   /api/resume/versions/upload
- GET    /api/resume/versions/compare?a=&b=
- GET    /api/resume/versions/<version_key>
- DELETE /api/resume/versions/<version_key>  (closes #44, preserved)

Same live-DB_PATH lookup idiom (from core import config as _config) as
the prior route extractions.

test_legacy_delete_consistency.py fixture: added one monkeypatch line so
core.config.DB_PATH is mutated to point at the test DB. Without it the
new blueprint would target the production default and the tests'd write
to a /tmp path mismatch.

═══ Server.py split complete ═══

Total transformation across 8 PRs (#60-#65 + this one):
  Before: backend/server.py = 905 LOC monolith
  After:  backend/server.py = 184 LOC (app init + blueprint registration +
                                       CORS hook + main entry)
  +       backend/core/config.py          (60 LOC, env constants + auth)
  +       backend/core/state.py           (72 LOC, State class + singleton)
  +       backend/core/scrape_loop.py     (146 LOC, BG scraper machinery)
  +       backend/routes/data_routes.py   (81 LOC, 5 read-only endpoints)
  +       backend/routes/scrape_routes.py (77 LOC, /api/scrape + status)
  +       backend/routes/profile_routes.py        (132 LOC, 5 routes)
  +       backend/routes/application_routes.py    (227 LOC, 6 routes)
  +       backend/routes/resume_version_routes.py (240 LOC, 5 routes)

Each file has a single concern. Future feature work no longer has to
load the whole 900-LOC monolith to find a route handler — git grep
hits the right module file directly.

Tests: 219 passed across all 8 PRs (no regressions, no new tests for
this PR since the routes themselves are byte-identical, just relocated).
